requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
streamlit==1.38.0
langchain
numpy==1.23.5
scipy==1.10.1
numba==0.56.4
librosa==0.9.2
pydub
moviepy
soundfile
tqdm
pyttsx3
python-dotenv
piper-tts
pillow
imageio
imageio-ffmpeg
scikit-image
opencv-python-headless==4.10.0.84